Working average calculator

See merge request comp3350-winter2024/KeyValuePairs-a02-8!8

+++ app/src/main/java/comp3350/student_echo/objects/InstructorReview.java
@@ -3,7 +3,7 @@ package comp3350.student_echo.objects;
 public class InstructorReview extends Review {
     Instructor instructor;
 
-    public InstructorReview(Instructor instructor, String comment, float overallRating, float difficultyRating) {
+    public InstructorReview(Instructor instructor, String comment, int overallRating, int difficultyRating) {
         super(comment, overallRating, difficultyRating);
         this.instructor = instructor;
     }
